Promoted a number of classes out of DepotRepository because they are used all across the package. Moved marshaller creation and management into the PersistenceContext because marshallers should be shared across a persistence context.
Also moved the actual "get a connection and do the query" code into the
persistence context which is sort of a nice separation. The repository and
friends now just put together Query and Modifier objects and pass them to the
context for execution. The context will eventually handle interaction with the
cache as well (exposing manual manipulation mechanisms as needed to the
DepotRepository).


git-svn-id: http://depot.googlecode.com/svn/trunk@26 80e7ffd8-b386-11dd-a77d-ab79ca5f3536