A whole bunch of revampery on the way to collection caching, including some basic stats reporting and query and cache logging.
PersistenceContext no longer sticks its nose so deeply into the business of
Query and Modifier. It just passes itself along and allows them to do any cache
activity they need during their normal execution. The one situation where we
keep our nose in their business is to allow a Query to return a result from the
cache before it is even invoked so that we can avoid requesting a database
connection if we won't ever need one.


git-svn-id: http://depot.googlecode.com/svn/trunk@343 80e7ffd8-b386-11dd-a77d-ab79ca5f3536