Great big revampage: no more strings anywhere. It's all ColumnExps all the way down (well, all the way to the public API anyway).
Revamped index creation while I was in there because that was one of the big
string users. Now you just put @Index on the field you want indexed, and if you
want a multi-column index you do things with a magical static method just like
we do for complex (function) indices.

@UniqueConstraint may still go away since it's basically exactly the same thing
as @Index(unique=true), so it's kind of pointless to support both.


git-svn-id: http://depot.googlecode.com/svn/trunk@376 80e7ffd8-b386-11dd-a77d-ab79ca5f3536