Charlie has convinced me that the database manipulation operations should be public and I should not force everyone who wants to use Depot to route all database accesses through custom public methods that hide all of the underlying Depot business behind application specific APIs.
I still think this is the best way to use Depot, but I acknowledge that there
are certain valid use cases (like generic code that operates on records in
arbitrary repositories) which are precluded by this scheme. So instead I'll
just strongly suggest that Depot be used in the way I think it should, and
frown sternly at anyone who deviates from The Way.


git-svn-id: http://depot.googlecode.com/svn/trunk@571 80e7ffd8-b386-11dd-a77d-ab79ca5f3536