Separate SeqImpl out of Sequence, hiding it away in our impl package tree. Make Sequence an interface.
We're going to change all of the places in Depot where we take a Collection as
an argument and take an Iterable instead. We invariably just turn the
Collection to an array, and we can just as well do that with an Iterable and be
more flexible about what we accept. Iterables.toArray has a fast path for
Iterable instances that also implement Collection, so we make SeqImpl implement
Collection under the hood even though we don't advertise that in the Sequence
interface because we don't want people using these as collections.

What we really want is something in between Iterable and Collection that knows
its size but otherwise does not allow mutation or containment checks.


git-svn-id: http://depot.googlecode.com/svn/trunk@580 80e7ffd8-b386-11dd-a77d-ab79ca5f3536