Well, we can at least write a unit test for this super special business.
However, I started making HSQL work a long time ago so that we could have real
unit tests that did real meaningful things, and Zell was kind enough to finish
making HSQL work, so we should be able to do some real end-to-end testing. I
just need to set up the necessary scaffolding. On the list!


git-svn-id: http://depot.googlecode.com/svn/trunk@595 80e7ffd8-b386-11dd-a77d-ab79ca5f3536