Wrangled Ivy into obtaining samskivert dependency from it's own "repository".
Why does the Java dependency management world have to suck so badly? Is it not
crazy to want projects to be able to publish their own releases and to make it
easy for other projects to depend on those published releases? Why do I have to
reconfigure the standard configuration to add new repository sources? Why
doesn't the IvyRep resolver work and do something sophisticated?

Why can't I just fucking put in my ivy.xml:

<dependency org="com.samskivert" name="samskivert" rev="1.0" conf="runtime"
            source="http://samskivert.googlecode.com/svn/releases"/>

and have Ivy do something smart like assume that there will be a standard
directory structure at the specified root and that it will contain the ivy.xml
and module.jar file that it needs, and just download them?

Instead I have to write a ton of fucking boilerplate and wire it all in via a
separate ivysettings.xml file, and then helpfully instruct anyone who uses my
fucking library that they too can have the pleasure of doing the same.


git-svn-id: http://depot.googlecode.com/svn/trunk@658 80e7ffd8-b386-11dd-a77d-ab79ca5f3536