Skip classes in GenRecordTask if they have @GeneratedValue on a field without @Id.  I'd rather stop the build, but other errors in class processing seem to lead to skipping, so I'm not rocking the boat.
If @GeneratedValue is on a field without an @Id, depot creates the necessary indices to generate the
value, but it doesn't actually do the generation on insertion.  Rather than making it work without a
good use case for it, I'm adding this check to keep people from stumbling into a broken design.



git-svn-id: http://depot.googlecode.com/svn/trunk@712 80e7ffd8-b386-11dd-a77d-ab79ca5f3536