Maven makes using version ranges too annoying by constantly checking for new versions with the repository.
Why, Maven, do you not offer an option to not go searching numerous sources on
the Internet every time I run a build? (And then timing out when some of those
sources are behind firewalls and I'm not running vpn at the moment.) Indeed,
why would that not be the default?


git-svn-id: http://depot.googlecode.com/svn/trunk@787 80e7ffd8-b386-11dd-a77d-ab79ca5f3536