Added OrderBy.Order.NULL. "order by null" can be used in a query with a "group by" in it to tell the database to suppress the overhead of sorting the results. (Supported in postgres and mysql.)
Changed the build visitors to have a whitelist of supported
Orders, and to NPE on a null value for the Order enum.

Note that previously "order by null" could be accomplished by passing
a null value to your OrderBy, but now the enum constant must be used.
I can change that if this disrupts anything.


git-svn-id: http://depot.googlecode.com/svn/trunk@891 80e7ffd8-b386-11dd-a77d-ab79ca5f3536