Added support for disabling schema migrations for a PersistenceContext.
This allows two separate services which share a database to assign "control" of
the schema to one of the services. The non-controlling service can be
configured to simply fail if new code is deployed to said service before new
code is deployed to the controlling service, rather than have the
non-controlling service migrate the schema out from under the controlling
service.

git-svn-id: http://depot.googlecode.com/svn/trunk@898 80e7ffd8-b386-11dd-a77d-ab79ca5f3536