Improve the semantics of @Transform handling.
@Transform is first sought on the field itself, which overrides any other
handling. Next it is sought on the class that represents the field (i.e. if the
field is type Foo we look for an @Transform annotation on class Foo). We avoid
doing this somewhat expensive search if the field has a type that we know
cannot have an @Transform annotation. If we've still found no @Transform
annotation (which is unfortunately the overwhelmingly common case), we then
create a stock marshaller.

git-svn-id: http://depot.googlecode.com/svn/trunk@908 80e7ffd8-b386-11dd-a77d-ab79ca5f3536